Implement boundry checking and some missing tests
I filled in the function onBoard(). All it does is take in an x and y value
and checks if it is on the Board. If it is True if Not False

All my tests were passing but there was some missing coverage in the move
function. Specifically it was not checking if movement worked in every
direction. I added three more tests for that. I also fixed a bug in
testBadPlace() as there was no way for it to pass.

Signed-off-by: slenderq <justinquaintance@gmail.com>